Turn off lights and use less power when idle
If the matrix hasn't changed in a while, turn off the display and the
LEDs on the BLE module.

We put all the rows HIGH and then just perform a simple single read
of all the columns each time we wake from the sleep mode (which is
really just a 15ms watchdog sleep).

The intent is to draw less power when the user is idle.

The wakeup keypress will typically get eaten because the wakeup
takes longer than the duration of the keypress; if that process
was a bit more smooth, then we could make the idle duration shorter.

I've set it to 10 minutes to trade away annoyance with reasonable
sounding power savings.