Fixup battery voltage reading, tweak power management a bit
This seems to correctly read the battery voltate level which is
attached to ADC channel 12.

Adjust the SSD1306 display code so that we render the status
information into an alternate buffer and only send it to the
device if it has changed.

While we're in there doing that, if the display hasn't changed
contents in the past 10 seconds, we just turn it off.  We'll turn
it back on again when we next flush the graphics.  This is
intended to extend the battery life a bit.