WIP: add quantum_console
This is an effort to make a quantum specific version of the hid_listen
tool.  This version of the code doesn't crash on disconnect on the mac,
and has slightly improve diagnostics on Linux.  It doesn't build on
Windows yet.

The goal is to be able to handle multiple different quantum keyboard
instances in a reasonable way.