pinnacle: Switch to absolute reporting
This seems to be more responsive than relative mode, but does mean
that we need to track some more state in the track class itself
so that we can generate the relative motion events for the os.

Clicking and dragging are what I'd call "functional" but finicky.
Should probably track a little bit more state to make sure the
experience is nice and smooth.